:safety_vest: Test: Productionize actions
