+++ packages/errors/src/actions/index.ts
@@ -7,3 +7,4 @@ export * from './LoadStateError.js'
 export * from './ScriptError.js'
 export * from './SetAccountError.js'
 export * from './ForkError.js'
+export * from './UnknownMethodError.js'
